Правки после ревью: Изменил отображение цены у NFT теперь это целочисленное значение Изменил структуру экрана каталога, кнопка сортировки теперь является NavifationItem'ом Измменил структуру экрана коллекции, поместил все элементы в ScrollView, добавил функцию динамического обновления высоты коллекции для корректного отображения на экране.

+++ FakeNFT/Scenes /Catalog/CollectionScreen/ViewModel/NFTCellModel.swift
@@ -11,7 +11,7 @@ struct NFTCellModel {
     let image: UIImage
     let rating: Int
     let name: String
-    let cost: Float
+    let cost: Int
     let isLike: Bool
     let inCart: Bool
 }
